add routes middleware
