+++ src/routes/index.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+router.get('/', (req, res) => {
+  const { url } = req;
+  res.json({ url });
+});
+
+module.exports = router;
